Convert LayoutRenderer to TS
